+++ src/test/java/roomescape/domain/reservation/ReservationRequestFixture.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import roomescape.application.dto.ReservationRequest;
